added few features

+++ controllers/feed.js
@@ -2,7 +2,7 @@ const {validationResult} = require('express-validator');
 const path = require('path');
 const fs = require('fs');
 const Post = require('../models/post');
-
+const User = require('../models/user')
 exports.getPosts = (req, res, next) => {
     const currentPage = parseInt(req.query.page) || 1;
     const perPage = 2;
@@ -44,26 +44,38 @@ exports.postPost = (req, res, next) => {
         throw error;
     }
     const title = req.body.title;
-    const imageUrl = 'images/' + req.file.filename
+    const imageUrl = 'images/' + req.file.filename;
     const content = req.body.content;
-    const creator = {
-        name: "Anonym"
-    };
+    const creator = req.userId;
     const post = new Post({title, imageUrl, content, creator});
-    post.save().then(result => {
-        res.status(201).json({
-            message: "Post crested successfully",
-            post: result
+    post.save()
+        .then(result => {
+            return User.findById(req.userId)
         })
-    }).catch(err => {
-        if (!err.statusCode) {
-            err.statusCode = 500
-        }
-        next(err)
-    });
+        .then(user => {
+            if (!user) {
+                const error = new Error('user not find')
+                error.statusCode = 404;
+                throw error;
+            }
+            user.posts.push(post)
+            return user.save()
+        })
+        .then(result => {
+            res.status(201).json({
+                message: "Post crested successfully",
+                post: post,
+            })
+        })
+        .catch(err => {
+            if (!err.statusCode) {
+                err.statusCode = 500
+            }
+            next(err)
+        });
 }
 
-exports.getPost = ((req, res, next) => {
+exports.getPost = (req, res, next) => {
     const postId = req.params.postId;
     Post.findById(postId)
         .then(post => {
@@ -84,14 +96,14 @@ exports.getPost = ((req, res, next) => {
             err.message = "Error while trying to get post from db";
             next(err)
         })
-})
+}
 
 exports.editPost = (req, res, next) => {
     const errors = validationResult(req);
     if (!errors.isEmpty()) {
-        const error = new Error("Data validation failed")
+        const error = new Error("Data validation failed");
         error.statusCode = 422;
-        throw error
+        throw error;
         return res.status(422).json({
             message: "Data validation failed",
             errors: errors.array()
@@ -100,46 +112,44 @@ exports.editPost = (req, res, next) => {
     const postId = req.params.postId,
         title = req.body.title,
         content = req.body.content;
-    let imageUrl = req.body.image;
-    if (req.file.filename) {
+    let imageUrl;
+    if (!!req.file) {
         imageUrl = 'images/' + req.file.filename;
-    }
-
-    if (!imageUrl) {
-        const error = new Error('please add some file');
-        error.statusCode = 422;
-        throw error;
     }
     Post.findById(postId)
         .then(p => {
             if (!p) {
+                console.log('DISPLAY ERROR ON INVALID PRODUCT')
                 const error = new Error('Cannot find post in the database');
                 error.statusCode = 404;
                 throw error;
             }
+            console.log(req.userId.toString() !== p.creator.toString())
+            if (req.userId.toString() !== p.creator.toString()) {
+                const error = new Error('not authorized');
+                if (!error.statusCode) {
+                    error.statusCode = 403;
+                }
+                throw error;
+            }
             p.title = title;
             p.content = content;
-            p.imageUrl = imageUrl;
+            p.imageUrl = imageUrl || p.imageUrl;
+            console.log(p.imageUrl)
             if (imageUrl !== p.imageUrl) {
                 clearImage(p.imageUrl);
             }
+            console.log('FINAL PRODUCT',p)
             return p.save()
-                .then(result => {
-                    res.status(200).json({message: 'Succesfully updated post.', post: result})
-                })
-                .catch(err => {
-                    if (!err.statusCode) {
-                        err.statusCode = 500;
-                    }
-                    err.message = "Error while trying to get post from db";
-                    next(err)
-                })
+        })
+        .then(result => {
+            res.status(200).json({message: 'Succesfully updated post.', post: result})
         })
         .catch(err => {
             if (!err.statusCode) {
                 err.statusCode = 500;
             }
-            err.message = "Error while trying to get post from db";
+            err.message = "Error while trying to get post from db EDITING";
             next(err)
         })
 }
@@ -153,12 +163,24 @@ exports.deletePost = (req, res, next) => {
                 error.statusCode = 404;
                 throw error;
             }
+            if (req.userId.toString() !== p.creator.toString()) {
+                const error = new Error('not authorized');
+                if (!error.statusCode) {
+                    error.statusCode = 403;
+                }
+                throw error;
+            }
             clearImage(p.imageUrl);
-            console.log(p);
             return Post.findByIdAndRemove(postId)
         })
         .then(result => {
-            console.log(result)
+            return User.findById(req.userId)
+        })
+        .then(result => {
+            user.posts.pull(postId)
+            return user.save()
+        })
+        .then(result => {
             res.status(200).json({message: 'deleted resource'})
         })
         .catch(err => {
